Attempting to move everything into a black box class, beginning to try to write BV algo with it.
